Refactor: Add PipelineManager
