+++ core/debugging.cpp
@@ -1,5 +1,6 @@
 #include <ThING/core.h>
 #include <ThING/extras/vulkanSupport.h>
+#include <cstring>
 
 void ProtoThiApp::populateDebugMessengerCreateInfo(VkDebugUtilsMessengerCreateInfoEXT& createInfo) {
     createInfo = {};
