+++ include/ThING/consts.h
@@ -1,6 +1,8 @@
 #pragma once
+#include <cstddef>
 #include <cstdint>
 
 inline const uint32_t WIDTH = 1200;
 inline const uint32_t HEIGHT = 800;
-inline const int MAX_FRAMES_IN_FLIGHT = 3;
+inline const size_t MAX_FRAMES_IN_FLIGHT = 3;
+inline const size_t CURRENT_PIPELINES = 2;
